logcat: Support searching through the message list.
This CL adds support for invoking a find dialog when
the logcat view is in focus. The dialog can be invoked
via Edit -> Find (Ctrl + F), or via the context menu.

The dialog provides a way to specify a search term and
allows the user to search forward or backward in the
list of messages.

Change-Id: I7e7c6b20a051c161f035b3b45aba5f119f2c11a9