Fix Cut & Paste in value files for Eclipse 4.x
This changeset works around the following bug in Eclipse 4.2:

* Open some arbitrary file, like main.xml.
* Open strings.xml for the first time. (If it has already been opened,
  switch to the graphical tab, close it, and start again).
* Place the caret in strings.xml and paste something from the
  clipboard. In Eclipse 4.x, strings.xml will not change, but the
  *other* file is edited to include the new paste!

This appears to be related to some sort of change in the defaults for
action bar registrations in Eclipse 4. The workaround is to explicitly
set these actions on each opened editor, which seems to fix the
problem in 4.2 (and things still work in 3.7).

Change-Id: I86035d7921102ff684adf319c3bf5f94ce416cb7