32745: Lack of warning of using duplicated IDs for element.
This changeset adds a validator to the assign/edit id dialog used in
the layout editor such that the user gets a warning if picking an id
which is already defined within the same layout.

Also cleans up the Rules API for adding a validator and makes the
generic resource validator handle both the case of requiring a unique
name and requiring an existing name.

Change-Id: Id9642c3bcd326f9734cf98c98f6799b67e11a4ae