Fix instrument ignoring extra arguments
Extra arguments passed to instrument are silently ignored. This could lead
to some confusion when instrument is invoked with extra arguments that are
expected to alter the instrumentation being run.

A common use case for this is when 'instrument' is invoked from monkeyrunner
to run all the tests in one class:

   device.instrument(pkg, { 'class':'com.example.test.MyTests' })

without this patch, the previous method will run all the tests in pkg.

Change-Id: I908d49422fe2755a1bcf562c2d040651b9691a6a