Workaround for Eclipse 4.x focus issue
In Eclipse 4, the system calls setFocus on the outline view *after* a
mouse click inside the outline has been processed. This ends up
calling setFocus on the treeviewer. If you clicked in the property
sheet to for example start editing a text property, this means the
focus loss will cause the property editor to abort editing again. The
net result is that you always need to click twice to start editing
properties if the Outline view does not already contain focus.

This changeset adds a workaround for this. It has two parts.

First, simply make the OutlinePage's setFocus() method do nothing if
it's called during a mouse click in the property sheet (which we can
track with a mouse listener).

The second part is trickier. The outline view itself (which is
provided by Eclipse and we have no hooks into it) will be called
first, and it calls setFocus() on its first child, which is a
composite, so it ends up searching down for the first focusable
component.  The workaround is to add our own widget as the first
child, and make it declare that it's focusable but make the setFocus
call a no-op. Second, make the widget invisible since we don't want
anything but the tree viewer in the outline.

Change-Id: Ib89943c410bce599723f59a4f202f339adab6327