Fix new project wizard's package validation
The wizard suggests "com.example." as a package prefix. However, the
validation would catch this as an invalid package and would just say
"Enter a package name". If you were editing "com.example.foo" this was
misleading. You now get a more specific warning when you've entered a
package name that has the example package as a prefix.

This CL also tweaks the focus handler which selects the prefix on
focus-entry to not include the separating dot in the prefix.

Change-Id: I659bb03515057eb5c760c51366869f4e0983f58a