Add @NonNull annotations to overridden method
Eclipse 4.2's new null analysis is useful, but (unfortunately)
requires nullness annotations to be repeated on methods overriding or
implementing interfaces where there are other nullness annotations.
This changeset adds this for the manifest merger.

(Eclipse Issue 381443 tracks a request for making the null analysis
not require this and simply pick up annotations from the overridden or
implemented method signature.)

Change-Id: I063b7713c5dbca642a8afd3231bbe4002ec130e1