Prevent zoom flicker
The layout editor preserves the zoom level of a file when you close
and reopen it later.  However, when a layout is opened for the first
time, there is a brief flicker as the layout is first shown at zoom
level 100%, and then briefly thereafter rezoomed to fit the available
canvas space.

The reason for this is that when the editor is first created, the size
of the layout editor is not known (SWT reports it as 0,0), so it can't
compute a correct zoom level. The earlier solution was to do a
Display.asyncExec() to rezoom soon.

This changeset improves upon this by observing that we don't need the
zoom to be recomputed until the canvas image is actually painted, so
the delayed zoom is queried lazily both by the paint routine as well
as the delayed setup runnable. This makes new layouts come up without
any zoom flicker.

Change-Id: I72a3a30fe40f26255951de50dcb8dcaeff44cbf7