Run lint in the layout editor after undo/redo
Lint already runs automatically after each edit in the layout editor.
However, the undo and redo actions in the layout editor are mapped
directly through to the XML editor's undo/redo actions (and when run
their edits trigger a model refresh). This meant that after an undo or
redo, lint would not run to update. If you for example delete a button
that has a lint error on it, then undo that edit, the button comes
back without the lint error.

This CL fixes this: it now wraps the Undo/Redo actions with an action
wrapper which first runs the delegated action, and then runs the edit
hooks (e.g. lint) afterwards.

Change-Id: I66431fd7989e1f3c0776563d159ca7b89fef8431