Populate ResourceFolder contents on demand
There are scenarios where clients call getFile(String name) on a
ResourceFolder before the ResourceFolder is updated after a file
change. For example, creating a new layout file can result in this
scenario, which causes a number of "Current file is not a match for
the given config." error messages in the layout editor.

This changeset fixes this such that if an existing file is not found
for a given name, it also checks the file system to see if the file
really exists, and if so, it processes it and adds it to the list.

It also makes some performance improvements. First, the code which
processes a resource folder would check each added file to see if
already existed in the list of resource files. This resource file list
was just stored as a list, so when adding files from a large folder
(such as the framework drawable-mdpi folder which has 1,200 files,
each successive added file would be compared against the previous n-1
files, O(n^2).) This changeset now stores a separate name map such
that it can quickly determine whether a file is already known.

Second, since some of the folders in the framework are large (and
their sizes are roughly known up front), it creates the data
structures with larger up front sizes (depending on the folder type)
to avoid expensive rehashing and array copying.

Change-Id: Ieaab955d7f165e790e8a8dccebf4d0878fde6fce