Add Activity Wizard
The "Android Application" wizard has an embedded "activity chooser"
page, and based on your selection the next page shows the
configuration page for that activity.

This changeset removes the two top level wizards for creating
activities ("New Blank Activity" and "New Master/Detail"), with a
single "New Activity" wizard whose first page is the activity chooser,
and second page is the same as it was under the two individual
wizards.

Other than making the menu simpler this also has the advantage that it
will pull in additional activities we register in the activities
folder (updated separately from ADT).

Change-Id: Iceef00178c4a0ceecc112db39f350285b7a223f7