Migrate Fragment and ListView metadata to the tools namespace
The persistent metadata storing AdapterView and Fragment bindings
predates the new tools namespace handled by aapt, and uses XML
comments. This changeset updates the code to using tools: attributes
instead.

Note that it doesn't migrate old settings; I could not find a way to
do that performantly (and in some cases the code which reads a setting
has no access to the document containing the metadata, only to parsed
XML nodes, so it cannot mutate it on demand to the new format.)
However, this shouldn't be a big deal; for fragments for example the
fragment message states what needs to be done to set up a binding.

This CL also fixes a couple of bugs around setting up these views
which should make switching views faster (it used to do the work
twice), and avoids a rare concurrent modification exception.

Change-Id: Id6a8a9a1649c1b9f6f5fc6a9fbc3a6e5b0512dd6