Provide readable device name rather than just serial number
There are a number of places where we need to display a device
name. Currently, DDMS shows the serial number + manufacturer/model
if available, but in other instances, we just show the serial
number.

This CL provides a getName() method in the IDevice interface so
that all the locations which require a human readable name can
use this.

The name is typically of the form manufacturer-model-serial.
If the manufacturer/model strings are not available, then those
fields are omitted, so it will show just the serial number.

Change-Id: I80ef7b48a4c7c8587042c1227c932251e18dba2e