Add format handling to the template wizards
This changeset adds a format attribute to the templates.  When the ADT
plugin reads a template file, and the format version is a higher
number than one it knows about, it rejects the templates and informs
the user to upgrade the plugin. This allows us to change the format of
the templates incompatibly in the future without worrying about users
with older versions of the plugin running into problems (because they
upgraded the tools but not ADT).

It also adds a revision number to each template which is used for the
opposite purpose: we can tell if a template is a slightly older
revision and then do some conditional handling.

Change-Id: If6c49012a1beca44c3a05b3e880ffde70294a57e