Fix emma for split projects.
The location of the .em file was wrong. Changing its location
means it gets put in the tested project so we need a way to know where
the output of the project is.

As we need to be better at supporting customized projects folder, we
introduce a new task that will be able to look for the customized
path of any projects. So far we only support the output folder
and the source folder.

Change-Id: Id8db82e254baae2088827c96b7719f94c92c927c