Fix convert8To4, convert8To5
See b/5680952 "Compilation warnings in etc1.cpp" for discussion.

This is a manual merge of an update that was made to
frameworks/native/opengl/libs/ETC1/etc1.cpp.

sdk/emulator/opengl/host/libs/Translator/GLcommon/etc1.cpp is an exact
copy of frameworks/native/opengl/libs/ETC1/etc1.cpp, so we might as well
keep the two versions in synch.

Bug: 5680952
Change-Id: Icf5d5ed2e7c5c79eb9677d210b1ff5fee507271d