Project wizard tweaks
Get rid of the ApplicationSkeleton page, and move the Create Custom
Icon checkbox up to the front page.  Also add a new "Mark as library
project" checkbox there, which allows to create a new library project.
Move the icon page up such that it appears before the activity
template pages. Fix a bug in minSdk handling.  And get rid of the
"Android Blank Project" wizard since it's now redundant with the
Android Application Project.

Change-Id: I54ea0bb740ad8078008a832799378694f37bf9e0