Wizard tweaks
Validate target location such that if the target
directory exists, the wizard complains.

Fix the icons for custom views and unit test projects
such that they have the "+" icon badge indicating that
something will be created. Update labels.

Replace the single theme defintion which references
DarkActionBar such that we have 3 styles, one for pre-11,
one for 11-13 and one for 14+ which define the relevant
light themes.

Remove the unused dimensions in blank projects.

Remove the (unused) "Include compatibility code" toggle.

Change-Id: I3d42c9544170c7a33754b0e410e9f1421d5c323b