Project template tweaks
Fix code which derives a project name from an application name such
that it skips invalid project name characters. For example, the
application name can be "I/O", but the project name should default to
just "IO".

Second, fix the validator code such that when it is validating a
template parameter constraint of "class", it validates it as a class
name, not a package name (so you cannot enter dots for example).

Third, replace the default project template icons with the standard
app icons (for when you disable Create Custom Launcher Icons).

Fourth, tweak the templates such that you don't have unused imports
when you create an activity without navigation, and remove unused
padding dimension resources.

Change-Id: I524608bb6347b4275338ec935e8de56bf3184c23