Use linear filtering for colorbuffer textures
Switch the min and mag filters for the ColorBuffer's texture object to
linear instead of nearest. This gives much better results when posting
a colorbuffer to a scaled framebuffer, which happens when the android
screen + emulator skin is too big to fit on the user's display.

The ColorBuffer's own texture object is only actually used as a
texture during post(); when used as a texture by the guest system
(e.g. by surfaceflinger composition) it's mapped through an EGLImage
to a guest-owned texture object, which has its own filter settings. So
we can just change the filter settings for the ColorBuffer texture
object at initialization without affecting anything else.

SDK Bug: 6721429

Change-Id: I8ec81125d076e0ed77a44f8b0dce412fa3cecabf