ADT fix: support Compatibility library in NPW
This fixes the NPW so that it supports the older
"Compatibility" library, not just the new "Support"
library. One issue is that the NPW prompts users to
update, but the code only checks for the support
lib in the new folder (SDK/extras/android/support).
During an upgrade the folder isn't changed and the
code fails to find it at the android/compatibility
location.

SDK Bug: 33859 and duplicates

Change-Id: I2a69d411e3875607859eb42ad2380ba069faf3b3