Compute completion documentation lazily, and add docs for ?-attrs
This changeset changes the content assist code to compute the
documentation popup text lazily (by providing our own
ICompletionProposal subclass which has enough data to compute it
lazily). This avoids the cost of formatting all the doc strings for
all the matches (there could be hundreds) and only doing it for the
currently selected item.

Second, this is used to look up lazily the documentation for attribute
strings such as ?android:attr/dividerHeight, which makes the recently
added code completion for these more useful.

Also fixes a typo in a field name in an unrelated class.

Change-Id: I12c37a5c91418d866d1d08f1401efdb07c1243de