Make the flag property editor editable
For "enum" properties (such as layout_width) we were using the
WindowBuilder combobox to let the user pick which of the enumeration
values to use. However, the enum values in Android often allow not
just the enumeration values, but other values too; for example, while
layout_width's enum values are match_parent and wrap_content, you can
also specify 42dp.

This changeset changes the property editor to use a similar property
editor to the one used for flag values; a free-form text field along
with field completion offering the available enum values. (We can't
simply just modify the property editor for enums to have an editable
combobox, since that property editor is a custom SWT component which
doesn't support editing.)

Change-Id: I1bff568233b4727d80ffbe7c17742177c6596381