Bump up available memory to lint cli
The lint cli can require a lot of memory when it's run on a large
directory full of project, since it accumulates all the errors as well
as source contents in order to be able to sort them and to emit source
code snippets in the HTML report. The default of 256 Mb isn't always
enough, so bump it up a bit.

For more, see http://code.google.com/p/android/issues/detail?id=34734

Change-Id: Ib7f0fd7a5a2d149221a43bb47490b44b46d39de0