Detect incorrect state lists
Lint already had a check for <selector> lists where it detects the
case where you have an empty child not in the last position; this will
match everything, and hide all subsequent states.

This changeset generalizes this check: It makes sure that no state is
fully matched by a previous state.

Change-Id: I43b80beb531e68008d82f6c620739ca179d0f8f8