Fix keybindings in various XML editors for Eclipse 4.x.
This changeset attempts to fix the issue with cut/copy/paste (and
other keybindings) not working properly in the non-layout XML
files. The symptom is that the action appears not to work, but is
actually applied to a different editor.

I haven't figured out why this happens in e4; after you switch focus
the editor action map should switch, but the fix here is to listen for
part changes, and manually reset the action bindings whenever this
happens.

This fixes issue 34630 (as well as several duplicates filed recently).

Change-Id: Ia1134c16ac3323a8a03f2c61182d810b7f7d2589