Fixed bug in creating already existing layout configs.
This fixes the issue where if you try to create a resource
qualified layout via the layout editor where one already
exists, ADT will overwrite the the contents of the existing
layout. Even worse, if the layout you try to create is the
one you're currently editing, it will replace it with an
empty file.

Change-Id: If3017c1c77d21cf99a4d77e62b46fb079997596e