35190: Lint doesn't detect illegal API usage in constructor
Constructors are not like methods: they are not inherited
into subclasses. The Lint API check needed some special
handling of constructors to avoid this, since it was
looking up inherited constructors and treating them as
accessible. In particular that meant it wouldn't flag
new IOException(Throwable) as requiring API 9, since its
superclass had a constructor which also took a Throwable
(with API 1).

Change-Id: I268752be506f12399ce4855d2cf7bec76012e4c2