SDK Manager fix: override in-memory sources when checking for updates.
An issue that showed up with the cache mechanism in Tools r20
is that users are not notified of tools updates unless they
clear their cache. The issue is that the SDK Manager first tries
to load an existing cache for a fast startup and then performs
a network check. However it doesn't clear the sources already
memory loaded in memory and thus skips the actual network check
and simply uses the first loaded result. This changes it so that
the network check actually happens even if there's a source
loaded in memory.

Change-Id: I3fde77b9aec8ccbd1cf66f79f99f8c5f4e84d900