Send full-size framebuffer to OnPost callback
When the emulator window has non-1.0 scaling, the scale is applied
when blitting the Android framebuffer to the host window. For the
OnPost callback, we were not only reading the image back from the
window (so post-scaling), we were also storing it in a 1.0-scaled
buffer. Now we read back from the Android framebuffer image, which is
always 1.0-scaled.

Change-Id: Ia9c974711a9a4b0b19f4b997f65ecc64481b4c6a