ndk: Allow empty path to NDK
There is no way to get out of the NDK preference page unless
a valid NDK is provided since Eclipse always pops up a message
saying that "the provided path is invalid".

This CL allows empty path to be a valid path and disables error
checking in such a case.

Change-Id: I4e5d5b975bd4629920f1b2750adec521c3a054a4