Improve handling of library projects in lint
This changeset improves the way lint handles library projects.

Until now, running lint on project "master" would also look up any
library projects referenced by the master project, and analyze the
library projects as well. This is necessary in order to correctly
compute unused resources for example, since it's possible for a
resource to be defined in one project and referenced in another.

However, while this behavior is desirable for users who partition
their code up into library projects, it has some serious problems for
users who are using a third party library project:

- Their lint output can be swamped with errors from the library which
  they have no control over.

- If for example the library provides translations into 60 languages,
  lint will use these 60 languages as the set of languages targeted by
  the application, and complain about all strings in the master
  project which are not translated into all the languages.

This changeset makes a key change to how library projects are
handled:

   (1) If you run lint on all projects (including the library projects),
       then there is no change from before.

   (2) If you run lint and specify just a project, then lint will
       continue to analyze the project as well as all its libraries,
       but will only report problems on the user-specified project.

The way this is done is by a new "report errors" attribute stored with
each project. All projects that are explicitly referenced on the
command line (or selected in the Eclipse UI), and all projects that
are found recursively if you specify a top level directory, all these
projects have their "report errors" flag set. Any remaining projects,
which would be those found through library project references, these
have their report errors flag cleared. And whenever lint is processing
errors, it will filter out errors for projects that are not reporting
errors.

This addresses issue
http://code.google.com/p/android/issues/detail?id=33847
as well as a number of other requests (in StackOverflow and elsewhere)
around the ability to filter errors in library projects.

Change-Id: I9d084f598c678ecf79cfe70d8ea7a84844333acc