Template support for minimum build target
This changeset adds a minBuildApi attribute to the <template>
element which can be used to require a minimum level of build
target (which can be higher than the template's required
minimum API). As with the minimum API, this can be specified
both on the root element, as well as on <option> elements.

Examples:
     name="New Blank Activity"
+    minBuildApi="12"
     description="Creates a new blank activity, with optional inner navigation.">
    <parameter id="navType" ...>
         <option id="tabs_pager" minApi="14">Tabs + Swipe</option>
+        <option id="pager_strip" minApi="14" minBuildApi="15">Swipe Views + Title Strip</option>
         <option id="dropdown" minApi="14">Dropdown</option>

Change-Id: I9d5b426bf2d440103a57f2a99ba6713cd6e7afaa