35588: @SuppressLint does not work on constructors in Java files
The AST-based check for whether an issue is suppressed did not take
constructors into account. This was because the AST node for
constructors (ConstructorDeclaration) does not extend the AST node for
method declarations (MethodDeclaration). This mistake was made in a
couple of other detectors as well.

(Also renamed test class which did not have the correct name (detector
class + "Test") so jump to test did not work.)

Change-Id: I0ecf99ab7d0357a03e300b7197ae84079f0ddbd1