Add support for theme selection in new projects
The project templates now contain a variable for controlling the
initial theme used with the project. This changeset adds support for
this in the New Project wizard, since it has custom UI code rather
than being data driven like most of the templates. In particular, it
conditionally adds the theme selector if it's available in the
template, and it handles validation based on the build target.
There's some code reorganization to be able to share UI code between
the New Template Page and the New Project Page.

Change-Id: I53e5602a0a983c4a1aea9ae7fc5cd1054ccd4e69