Clear lint warnings when the min/target SDK levels change
When you edit Java files, we automatically mark API errors - calling
an API which requires a higher minSdkVersion than the current one
specified in the manifest.

If you then switch to the maniest editor, change the value to the new
desired API level, save the file and come back, the API errors are
still marked in the Java file.

The reason this happens is that the API checker runs incrementally on
Java files, but it's affected by manifest values. This has been
reported by multiple users.

We can't easily perform incremental analysis in this case, and we
don't want to kick off a global rescan of lint, but to avoid
confusion, we can *clear* all lint markers when the minSdkVersion or
targetSdkVersion change. This will make confusing errors go away, and
incremental warnings will be shown when the user next edits the file.
It has the downside that a user who runs Lint on the whole project,
then edits the SDK version and saves the file will suddenly find all
the lint errors cleared, but they can just re-run lint, which is
hopefully a smaller problem than seeing confusing error messages (at
least two users reasonably did not realize that these were stale
errors and thought there was a remaining problem).

Finally, tweak the ManifestInfo class a bit; add nullness annotations,
fix an incorrect usage from the project wizard, fix bug in version
name computation and add a getter to return the code name of the
version name.

Change-Id: I5872733d2f12a4950264d08e689c7c0c025c586e