Include additional metadata in the XML report
This CL adds additional metadata into the lint XML report, such as the
full issue explanation, the issue priority, the issue summary, the
additional information URL, and the error line.

This allows for example the Jenkins plugin to provide all this
information for a build report without having to bundle a (possibly
out of date) snapshot of lint's metadata the way it does right now.

Change-Id: Iaf674ffa6bbef318ce3f636404d72c74de4b1c50