Fullscreen activity template.
This adds an activity template for a fullscreen activity, such as a video
player. The content itself is placeholder content, but this handles the
important bits like showing and hiding system navigation and handling
orientation changes manually (correctly, as of Android 3.2 with the
screenSize config flag).

This adds quite a bit of utility code (SystemUiHider). In the future
we should move this out of the template and into the support library.

Change-Id: I9c9c4c4699a3cec9659171d84bdbd23a5bd41178