35875: Lint too strict about translating strings by default.
This changeset updates the issue explanation for the translation
detector to explain the translatable=false and donottranslate.xml
mechanisms to handle non-translatable strings.

It adds a quickfix for missing translation items to set the
translatable attribute to false.

It also makes lint warn where you're using translatable=false
on strings in a locale-folder, as well as when you're translating
a string defined as translatable=false in the base folder.

And finally it also bumps up the severity of the extra translations
issue as justified in issue 35875.

Change-Id: I7539464b234b0a4b444bf9f188ce5b819f962430