Update lint unit test golden files to contain more details
The lint golden files was using a custom unit test reporter which
wrote out the test strings in alphabetical order. However, this output
was missing some information such as the actual source lines
containing the error and the particular error range underlined, which
the CLI prints.

This changeset replaces the unit test reporter with the CLI one, which
not only has the benefit of providing more details in the unit tests
(which for example shows that bytecode based issues correctly identify
the right source level range), but it also tests the CLI reporter
itself such as its sorting.

Other than a minor fix to handling \r's in the CLI error output, this
changeset does not contain any changes to any of the lint detectors,
it just updates the golden files.

Change-Id: Iae8f13a47efe6ba5e5cfd84af0dd548f78513879