Asset studio fixes
When generating actionbar icons from clipart, do not strip surrounding
space, and do not add extra padding. The clipart images already
contain baked in padding suitable for action bar icons. This CL also
makes sure the UI disables the corresponding options in this mode.

It also renames and moves to the bottom the legacy menu and tab icon
generators, and fixes a file resource leak in the image loading code.

Change-Id: I8a0dd61c97862206cdc71dc591a207a0b6a050f8