Improvements to relative layout move and delete operations
This changeset improves the way the RelativeLayout editing support in
the layout editor handles deletions and moves.

First, during a move, if the move is simply within the same layout,
then the layout constraints are left alone such that if you for
example have

      A
      v
      B < C < D

and you move B up to be next to A, you end up with

      A < B < C < D

(It will however remove cycles if the move would result in them.)

Second, it now handles deletion better where deleting a view will
cause all references to any deleted views to be replaced by transitive
constraints.

For example, if you hve

     A < B < C < D

and you delete B and C, you end up with

     A < D

Change-Id: Icb9d3552e60aee20192f7941fe52be71ba52557f