uiautomator: Misc cleanups
This CL performs a bunch of minor cleanups to allow reusing the
UiAutomatorView from within an Eclipse view/editor. In particular:
 - The model is not a singleton anymore. It can be constructed
   based on an XML file, and is stored within the view.
 - References to the view from the model have all been removed.
 - All the view code from the application window class has been
   moved into a separate UiAutomatorView class.

There is no new functionality, and from the UI perspective, there
is only one change: The are now two toolbars, a global application
level toolbar for global actions (open file, capture screenshot),
and a toolbar above the view hierarchy that allows setting a couple
of tree viewer options.

Change-Id: Ib4fe98426c2f83de233091c23b080de4f9b7449a